+++ integration_test/gesture_and_stress_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:glass_bottom_bar_ui/app.dart';
-import 'package:glass_bottom_bar_ui/glass_morphism_ui.dart';
+import 'package:liquid_glass_ui/app.dart';
+import 'package:liquid_glass_ui/glass_morphism_ui.dart';
 import 'package:integration_test/integration_test.dart';
 
 void main() {