+++ lib/glass_ui_kit.dart
@@ -1,15 +1,16 @@
 library glass_ui_kit;
 
 // Core Contracts & Architecture
-export 'package:glass_bottom_bar_ui/core/glass_navigation_controller.dart';
-export 'package:glass_bottom_bar_ui/core/glass_scope.dart';
-export 'package:glass_bottom_bar_ui/core/glass_ui_contracts.dart';
+export 'package:liquid_glass_ui/core/glass_navigation_controller.dart';
+export 'package:liquid_glass_ui/core/glass_scope.dart';
+export 'package:liquid_glass_ui/core/glass_ui_contracts.dart';
 
 // Reusable Glass UI Components & Extensions
-export 'package:glass_bottom_bar_ui/src/glass_container.dart';
-export 'package:glass_bottom_bar_ui/src/glass_ui_style.dart';
-export 'package:glass_bottom_bar_ui/src/glass_widget_extension.dart';
+export 'package:liquid_glass_ui/src/glass_app_bar.dart';
+export 'package:liquid_glass_ui/src/glass_container.dart';
+export 'package:liquid_glass_ui/src/glass_ui_style.dart';
+export 'package:liquid_glass_ui/src/glass_widget_extension.dart';
 
 // UI Screens & Models
-export 'package:glass_bottom_bar_ui/glass_morphism_ui.dart';
-export 'package:glass_bottom_bar_ui/liquid_glass_ui.dart';
+export 'package:liquid_glass_ui/glass_morphism_ui.dart';
+export 'package:liquid_glass_ui/liquid_glass_ui.dart';