+++ test/unit_and_painter_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:glass_bottom_bar_ui/liquid_glass_ui.dart';
+import 'package:liquid_glass_ui/liquid_glass_ui.dart';
 
 void main() {
   group('BottomNavigationItem Unit Tests', () {